pyproject.toml
[project]
name = "jiratcgeneratorandplaywrightcodegenerator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "anthropic>=0.75.0",
    "chromadb>=1.0.0",
    "crewai>=0.1.24",
    "cryptography>=43.0.0",
    "langchain-anthropic>=1.3.0",
    "langchain-openai>=1.1.6",
    "litellm>=1.80.0",
    "openai>=1.109.1",
    "openpyxl>=3.1.0",
    "pandas>=2.3.3",
    "playwright>=1.40.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
]
